pyproject.toml
[project]
name = "resume-insight-ai"
version = "0.1.0"
description = "AI-powered resume analysis and comparison app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.77.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
